hv: nested: Enable Preemption-Timer for hv console (vmcs01)
This patch enables vmx preemption timer for debugging purpose.

In nested environment, hv will lose all timer interrupts due to LAPIC
passthrough and therefore the hv console is inaccessible. This patch
enables the preemption timer as an alternative to drive the hv console.

The preemption timer should be enabled on vmcs01 and vmcs02. This patch
enables it on vmcs01.

Signed-off-by: Yifan Liu <yifan1.liu@intel.com>